+++ bootstrap3/renderers.py
@@ -17,11 +17,12 @@ from django.forms.formsets import BaseFormSet
 from django.utils.html import conditional_escape, escape, strip_tags
 from django.utils.safestring import mark_safe
 
-from .bootstrap import get_bootstrap_setting
+from .bootstrap import get_bootstrap_setting, DJANGO_MAX_18
 from .exceptions import BootstrapError
 from .forms import (
     render_form, render_field, render_label, render_form_group,
-    is_widget_with_placeholder, FORM_GROUP_CLASS
+    is_widget_with_placeholder, FORM_GROUP_CLASS,
+    is_widget_required_attribute
 )
 from .text import text_value
 from .utils import add_css_class, render_template_file
@@ -41,6 +42,11 @@ class BaseRenderer(object):
         self.show_help = kwargs.get('show_help', True)
         self.show_label = kwargs.get('show_label', True)
         self.exclude = kwargs.get('exclude', '')
+
+        # Only relevant if DJANGO_MAX_18
+        self.set_required = kwargs.get('set_required', True)
+        self.set_disabled = kwargs.get('set_disabled', False)
+
         self.set_placeholder = kwargs.get('set_placeholder', True)
         self.size = self.parse_size(kwargs.get('size', ''))
         self.horizontal_label_class = kwargs.get(
@@ -106,6 +112,8 @@ class FormsetRenderer(BaseRenderer):
                 show_label=self.show_label,
                 show_help=self.show_help,
                 exclude=self.exclude,
+                set_required=self.set_required,  # Only relevant if DJANGO_MAX_18
+                set_disabled=self.set_disabled,  # Only relevant if DJANGO_MAX_18
                 set_placeholder=self.set_placeholder,
                 size=self.size,
                 horizontal_label_class=self.horizontal_label_class,
@@ -148,6 +156,11 @@ class FormRenderer(BaseRenderer):
                 'Parameter "form" should contain a valid Django Form.')
         self.form = form
         super(FormRenderer, self).__init__(*args, **kwargs)
+
+        # Handle form.empty_permitted
+        if DJANGO_MAX_18 and self.form.empty_permitted:
+            self.set_required = False
+
         self.error_css_class = kwargs.get('error_css_class', None)
         self.required_css_class = kwargs.get('required_css_class', None)
         self.bound_css_class = kwargs.get('bound_css_class', None)
@@ -164,6 +177,8 @@ class FormRenderer(BaseRenderer):
                 show_label=self.show_label,
                 show_help=self.show_help,
                 exclude=self.exclude,
+                set_required=self.set_required,  # Only relevant if DJANGO_MAX_18
+                set_disabled=self.set_disabled,  # Only relevant if DJANGO_MAX_18
                 set_placeholder=self.set_placeholder,
                 size=self.size,
                 horizontal_label_class=self.horizontal_label_class,
@@ -280,6 +295,12 @@ class FieldRenderer(BaseRenderer):
         if self.field.form.empty_permitted:
             self.required_css_class = ''
 
+        # Special case to support Django 1.8 required / disabled
+        if DJANGO_MAX_18:
+            if self.field.form.empty_permitted:
+                self.set_required = False
+            self.set_disabled = kwargs.get('set_disabled', False)
+
     def restore_widget_attrs(self):
         self.widget.attrs = self.initial_attrs.copy()
 
@@ -313,6 +334,24 @@ class FieldRenderer(BaseRenderer):
                 escape(strip_tags(self.field_help))
             )
 
+    def add_required_attrs(self, widget=None):
+        """
+        Only relevant if DJANGO_MAX_18
+        """
+        if widget is None:
+            widget = self.widget
+        if self.set_required and is_widget_required_attribute(widget):
+            widget.attrs['required'] = 'required'
+
+    def add_disabled_attrs(self, widget=None):
+        """
+        Only relevant if DJANGO_MAX_18
+        """
+        if widget is None:
+            widget = self.widget
+        if self.set_disabled:
+            widget.attrs['disabled'] = 'disabled'
+
     def add_widget_attrs(self):
         if self.is_multi_widget:
             widgets = self.widget.widgets
@@ -322,6 +361,9 @@ class FieldRenderer(BaseRenderer):
             self.add_class_attrs(widget)
             self.add_placeholder_attrs(widget)
             self.add_help_attrs(widget)
+            if DJANGO_MAX_18:
+                self.add_required_attrs(widget)
+                self.add_disabled_attrs(widget)
 
     def list_to_class(self, html, klass):
         classes = add_css_class(klass, self.get_size_class())