+++ demo/demo/settings.py
@@ -157,6 +157,7 @@ LOGGING = {
 
 # Settings for django-bootstrap3
 BOOTSTRAP3 = {
+    'set_required': False,  # For Django <= 1.8 only
     'error_css_class': 'bootstrap3-error',
     'required_css_class': 'bootstrap3-required',
     'javascript_in_head': True,