+++ HISTORY.rst
@@ -6,8 +6,7 @@ History
 Develop
 +++++++
 
-* **BREAKING** Remove everything to do with setting HTML `required` attribute (#337)
-* **BREAKING** Remove everything to do with setting HTML `disabled` attribute (#345)
+* **BREAKING** For Django > 1.8 Remove everything to do with setting HTML attributes `required` (#337) and `disabled` (#345)
 * Add `id` parameter to bootstrap_button (#214)
 * Add `set_placeholder` to field and form renderers (#339, thanks @predatell)
 * Default button type to `btn-default`